WIP: changed realmSpawn to take a TaskFuncPtr
This builds successfully.  Needs a bunch of code cleanup. The Function realmSpawn now takes a Realm::Processor::TaskFuncPtr as an argument and creates the CodeDescriptor directly from it.  Yes, this would be a hard thing to pass in C, but it could be done.  The reasoning is that I can codegen a funtion poiter of appropriate type within LLVM regardless of the task's actual funciton type.